==> 0_Augmentation.ipynb <==
import os

from joblib import Parallel, delayed
from tqdm import tqdm_notebook as tqdm
import pickle
import pandas as pd

import jams
import numpy as np

import muda

def root(x):
    return os.path.splitext(os.path.basename(x))[0]

ANNOS = jams.util.find_with_extension('chordlab', 'jams')
AUDIO = jams.util.find_with_extension('data', 'flac')

# Make sure there are the same number of files
assert len(AUDIO) == len(ANNOS)

# And that they're in agreement
assert all([root(_1) == root(_2) for (_1, _2) in zip(AUDIO, ANNOS)])

def augment(afile, jfile, deformer, outpath):
    
    jam = muda.load_jam_audio(jfile, afile)
    
    base = root(afile)
    outfile = os.path.join(outpath, base)
    
    for i, jam_out in enumerate(deformer.transform(jam)):
        muda.save('{}.{}.flac'.format(outfile, i),
                  '{}.{}.jams'.format(outfile, i),
                 jam_out, strict=False)

# Create the augmentation engine for 1-6 half-note
pitcher = muda.deformers.PitchShift(n_semitones=[-1, 1, -2, 2, -3, 3, -4, 4, -5, 5, -6, 6])

with open('working/chords/muda.pkl', 'wb') as fd:
    pickle.dump(pitcher, fd)

OUT_DIR = 'working/augmentation/'

for (aud, jam) in zip(AUDIO, ANNOS):

    augment(aud, jam, pitcher, OUT_DIR) 

# # Build a pump
# sr = 44100
# hop_length = 4096

# p_feature = pumpp.feature.CQTMag(name='cqt', sr=sr, hop_length=hop_length, log=True, conv='tf', n_octaves=6)
# p_chord_tag = pumpp.task.ChordTagTransformer(name='chord_tag', sr=sr, hop_length=hop_length, sparse=True)
# p_chord_struct = pumpp.task.ChordTransformer(name='chord_struct', sr=sr, hop_length=hop_length, sparse=True)

# pump = pumpp.Pump(p_feature, p_chord_tag, p_chord_struct)

# # Save the pump

# with open('/home/bmcfee/working/chords/pump.pkl', 'wb') as fd:
#     pickle.dump(pump, fd)
