==> 1_DataDump.ipynb <==
import os
from glob import glob

from joblib import Parallel, delayed
import tqdm.notebook as tqdm
import pickle
import pandas as pd

import pumpp
import jams
import numpy as np

def root(x):
    return os.path.splitext(os.path.basename(x))[0]

ANNOS = jams.util.find_with_extension('chordlab', 'jams')
AUDIO = jams.util.find_with_extension('data', 'flac')

# Make sure there are the same number of files
assert len(AUDIO) == len(ANNOS)

# And that they're in agreement
assert all([root(_1) == root(_2) for (_1, _2) in zip(AUDIO, ANNOS)])

# Build a pump
sr = 44100
hop_length = 4096

p_feature = pumpp.feature.CQTMag(name='cqt', sr=sr, hop_length=hop_length, log=True, conv='tf', n_octaves=6)
p_chord_tag = pumpp.task.ChordTagTransformer(name='chord_tag', sr=sr, hop_length=hop_length, sparse=True)
p_chord_struct = pumpp.task.ChordTransformer(name='chord_struct', sr=sr, hop_length=hop_length, sparse=True)

pump = pumpp.Pump(p_feature, p_chord_tag, p_chord_struct)

# Save the pump

with open('working/chords/pump.pkl', 'wb') as fd:
    pickle.dump(pump, fd)

def convert(aud, jam, pump, outdir):
    
    data = pump.transform(aud, jam)
    
    fname = os.path.extsep.join([root(aud), 'npz'])
    
    np.savez(os.path.join(outdir, fname), **data)

OUTDIR = 'working/chords/pump/'

Parallel(n_jobs=20, verbose=10)(delayed(convert)(aud, jam, pump, OUTDIR) for (aud, jam) in zip(AUDIO, ANNOS));

import shutil

dest = 'working/chords/augmentation/'

ANNOS_A = jams.util.find_with_extension('chordlab', 'jams')
for anno in ANNOS_A:
    dest_path = dest + root(anno) + '.jams'
    print(dest_path)
    os.remove(dest_path)


# ANNOS_A = jams.util.find_with_extension('chordlab', 'jams')
# for anno in ANNOS_A:
#     print(root(anno))
#     for i in range(12):
#         dest_path = dest + root(anno) + '.' + str(i) + '.jams'
#         print()
#         shutil.copy(anno, dest_path)

# Augmented data
from glob import glob

AUDIO_A = jams.util.find_with_extension('working/chords/augmentation/', 'flac')
ANNOS_A = jams.util.find_with_extension('working/chords/augmentation/', 'jams')

# Make sure there are the same number of files
assert len(AUDIO_A) == len(ANNOS_A)

# And that they're in agreement
assert all([root(_1) == root(_2) for (_1, _2) in zip(AUDIO_A, ANNOS_A)])

Parallel(n_jobs=20, verbose=10)(delayed(convert)(aud, jam, pump, OUTDIR) for (aud, jam) in zip(AUDIO_A, ANNOS_A));

index = pd.Series()

null_artist = 0

for ann in ANNOS:
    J = jams.load(ann, validate=False)
    if not J.file_metadata.artist:
        artist = 'artist_{:05d}'.format(null_artist)
        null_artist += 1
    else:
        artist = J.file_metadata.artist
        
    index[root(ann)] = artist

index.to_json('working/chords/artist_index.json')
